% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defmt_pct}
\alias{defmt_pct}
\title{Deformat percent}
\usage{
defmt_pct(pct)
}
\arguments{
\item{pct}{string eager to remove percent}
}
\value{
Numeric value without percent
}
\description{
Deformat percent
}
\examples{
defmt_pct("10.0\%")
defmt_pct(c("10.0\%", "(11.2\%)"))
}
