% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sl_summary.R
\name{format_trt_compliance}
\alias{format_trt_compliance}
\title{Format Treatment Compliance Analysis}
\usage{
format_trt_compliance(
  outdata,
  display_col = c("n", "prop", "total"),
  digits_prop = 1,
  display_stat = c("mean", "sd", "se", "median", "q1 to q3", "range")
)

format_trt_compliance(
  outdata,
  display_col = c("n", "prop", "total"),
  digits_prop = 1,
  display_stat = c("mean", "sd", "se", "median", "q1 to q3", "range")
)
}
\arguments{
\item{outdata}{A metadata object created by \code{\link[=prepare_sl_summary]{prepare_sl_summary()}}.}

\item{display_col}{Column wants to display on the table.
The term could be selected from \code{c("n", "prop", "total")}.}

\item{digits_prop}{Number of digits for proportion columns.}

\item{display_stat}{A vector of statistics term name.
The term name could be selected from
\code{c("mean", "sd", "se", "median", "q1 to q3", "range", "q1", "q3", "min", "max")}.}
}
\value{
A list of analysis raw datasets.

A list of analysis raw datasets.
}
\description{
Format Treatment Compliance Analysis

Format Treatment Compliance Analysis
}
\examples{
meta <- metalite::meta_adam(
  population = metalite_sl_adsl,
  observation = metalite_sl_adsl
) |>
  metalite::define_plan(metalite::plan(
    analysis = "trt_compliance", population = "apat",
    observation = "apat", parameter = "comp8;comp16"
  )) |>
  metalite::define_population(
    name = "apat", group = "TRTA", subset = SAFFL == "Y"
  ) |>
  metalite::define_parameter(
    name = "comp8", var = "COMP8FL", label = "Compliance (Week 8)"
  ) |>
  metalite::define_parameter(
    name = "comp16", var = "COMP16FL", label = "Compliance (Week 16)"
  ) |>
  metalite::define_analysis(
    name = "trt_compliance", title = "Summary of Treatment Compliance"
  ) |>
  metalite::meta_build()

meta |>
  prepare_trt_compliance(parameter = "comp8;comp16") |>
  format_trt_compliance()
meta <- metalite::meta_adam(
  population = metalite_sl_adsl,
  observation = metalite_sl_adsl
) |>
  metalite::define_plan(metalite::plan(
    analysis = "trt_compliance", population = "apat",
    observation = "apat", parameter = "comp8;comp16"
  )) |>
  metalite::define_population(
    name = "apat", group = "TRTA", subset = SAFFL == "Y"
  ) |>
  metalite::define_parameter(
    name = "comp8", var = "COMP8FL", label = "Compliance (Week 8)"
  ) |>
  metalite::define_parameter(
    name = "comp16", var = "COMP16FL", label = "Compliance (Week 16)"
  ) |>
  metalite::define_analysis(
    name = "trt_compliance", title = "Summary of Treatment Compliance"
  ) |>
  metalite::meta_build()

meta |>
  prepare_trt_compliance(population = "apat", parameter = "comp8;comp16") |>
  format_trt_compliance()
}
