% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_sl_example.R
\name{meta_sl_example}
\alias{meta_sl_example}
\title{Create an example \code{meta_sl_example} object}
\usage{
meta_sl_example()
}
\value{
A metadata object.
}
\description{
This function is only for illustration purpose.
r2rtf is required.
}
\examples{
meta_sl_example()
}
