% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_trt_compliance.R
\name{prepare_trt_compliance}
\alias{prepare_trt_compliance}
\title{Prepare data for treatment compliance table}
\usage{
prepare_trt_compliance(
  meta,
  analysis = "trt_compliance",
  population = meta$plan[meta$plan$analysis == analysis, ]$population,
  parameter = paste(meta$plan[meta$plan$analysis == analysis, ]$parameter, collapse =
    ";")
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{analysis}{A character value of analysis term name.
The term name is used as key to link information.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare data for treatment compliance table
}
\examples{
meta <- metalite::meta_adam(
  population = metalite_sl_adsl,
  observation = metalite_sl_adsl
) |>
  metalite::define_plan(metalite::plan(
    analysis = "trt_compliance",
    population = "apat",
    observation = "apat",
    parameter = "comp8;comp16;comp24"
  )) |>
  metalite::define_population(
    name = "apat",
    group = "TRTA",
    subset = SAFFL == "Y"
  ) |>
  metalite::define_parameter(
    name = "comp8",
    var = "COMP8FL",
    label = "Compliance (Week 8)"
  ) |>
  metalite::define_parameter(
    name = "comp16",
    var = "COMP16FL",
    label = "Compliance (Week 16)"
  ) |>
  metalite::define_parameter(
    name = "comp24",
    var = "COMP24FL",
    label = "Compliance (Week 24)"
  ) |>
  metalite::define_analysis(
    name = "trt_compliance",
    title = "Summary of Treatment Compliance"
  ) |>
  metalite::meta_build()

meta |> prepare_trt_compliance()
}
