% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_base_char.R
\name{format_base_char}
\alias{format_base_char}
\title{Prepare data for baseline characteristic table}
\usage{
format_base_char(
  outdata,
  display_col = c("n", "prop", "total"),
  digits_prop = 1,
  display_stat = c("mean", "sd", "se", "median", "q1 to q3", "range")
)
}
\arguments{
\item{outdata}{A metadata object created by \code{\link[=prepare_base_char]{prepare_base_char()}}.}

\item{display_col}{Column wants to display on the table. Term could be selected from \code{c("n", "prop", "total")}.}

\item{digits_prop}{Number of digits for proportion columns.}

\item{display_stat}{A vector of statistics term name.
The term name could be selected from c("mean", "sd", "se", "median", "q1 to q3", "range", "q1", "q3", "min", "max").}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare data for baseline characteristic table
}
\examples{
meta <- meta_sl_example()

meta |>
  prepare_base_char(parameter = "age;gender") |>
  format_base_char()
}
