% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_base_char.R
\name{prepare_base_char}
\alias{prepare_base_char}
\title{Prepare data for baseline characteristic table}
\usage{
prepare_base_char(
  meta,
  population = "apat",
  observation = "wk12",
  analysis = "base_char",
  parameter = paste(names(meta$parameter), collapse = ";")
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{analysis}{A character value of analysis term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare data for baseline characteristic table
}
\examples{
meta <- meta_sl_example()

meta |>
  prepare_base_char()
}
