% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_disposition.R
\name{prepare_disposition}
\alias{prepare_disposition}
\title{Prepare data for disposition table}
\usage{
prepare_disposition(
  meta,
  population = "apat",
  observation = "apat",
  analysis = "disp",
  parameter = "disposition;medical-disposition"
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{analysis}{A character value of analysis term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare data for disposition table
}
\examples{
meta <- meta_sl_example()

meta |>
  prepare_base_char()
}
