% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_exp_duration.R
\name{prepare_exp_duration}
\alias{prepare_exp_duration}
\title{Prepare data for exposure duration table}
\usage{
prepare_exp_duration(
  meta,
  analysis = "exp_dur",
  population = meta$plan[meta$plan$analysis == analysis, ]$population,
  parameter = paste(meta$plan[meta$plan$analysis == analysis, ]$parameter, collapse =
    ";")
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{analysis}{A character value of analysis term name.
The term name is used as key to link information.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare data for exposure duration table
}
\examples{
meta <- meta_sl_exposure_example()
meta |> prepare_exp_duration()
}
