% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_trt_compliance.R
\name{prepare_trt_compliance}
\alias{prepare_trt_compliance}
\title{Prepare data for baseline characteristic table}
\usage{
prepare_trt_compliance(
  meta,
  analysis = "trt_compliance",
  population = meta$plan[meta$plan$analysis == analysis, ]$population,
  observation = meta$plan[meta$plan$analysis == analysis, ]$observation,
  parameter = paste(meta$plan[meta$plan$analysis == analysis, ]$parameter, collapse =
    ";"),
  display_total = TRUE
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{analysis}{A character value of analysis term name.
The term name is used as key to link information.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}

\item{display_total}{A logic value of displaying the total group.}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare data for baseline characteristic table
}
\examples{
meta <- meta_sl_example(type="comp")
meta |> prepare_trt_compliance()
}
