% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_base_char.R
\name{rtf_disposition}
\alias{rtf_disposition}
\title{Format Disposition Analysis}
\usage{
rtf_disposition(
  outdata,
  source,
  col_rel_width = NULL,
  text_font_size = 9,
  orientation = "portrait",
  footnotes = NULL,
  title = NULL,
  path_outdata = NULL,
  path_outtable = NULL
)
}
\arguments{
\item{outdata}{A metadata object created by \code{\link[=prepare_sl_summary]{prepare_sl_summary()}}.}
}
\value{
A list of analysis raw datasets.
}
\description{
Format Disposition Analysis
}
\examples{
meta <- meta_sl_example()

meta |>
  prepare_disposition(parameter = "disposition;medical-disposition") |>
  format_disposition()
  rtf_disposition(
    source = "Source: [CDISCpilot: adam-adsl]",
    path_outdata = tempfile(fileext = ".Rdata"),
    path_outtable = tempfile(fileext = ".rtf")
  )
}
