% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_base_char.R
\name{rtf_trt_compliance}
\alias{rtf_trt_compliance}
\title{Format Treatment Compliance Analysis}
\usage{
rtf_trt_compliance(
  outdata,
  source,
  col_rel_width = NULL,
  text_font_size = 9,
  orientation = "portrait",
  footnotes = NULL,
  title = NULL,
  path_outdata = NULL,
  path_outtable = NULL
)
}
\arguments{
\item{outdata}{A metadata object created by \code{\link[=prepare_sl_summary]{prepare_sl_summary()}}.}
}
\value{
A list of analysis raw datasets.
}
\description{
Format Treatment Compliance Analysis
}
\examples{
meta <- meta_sl_example()

meta |>
  prepare_trt_compliance(parameter = "comp8;comp16") |>
  format_trt_compliance() |>
  rtf_trt_compliance(
    source = "Source: [CDISCpilot: adam-adsl]",
    path_outdata = tempfile(fileext = ".Rdata"),
    path_outtable = tempfile(fileext = ".rtf")
  )
}
